Namn på filet + PopupMenuButton istället för ShowDialog
Filernas namn ska nu va i enlighet med flutters regler samt bytte jag från showdialog till popupmenubutton för att det tydligen va en grej och funkar betydligt bättre
